+++ clang/test/SemaCXX/thread-safety-annotations.h
@@ -31,8 +31,10 @@
 // Capabilities only
 #define EXCLUSIVE_UNLOCK_FUNCTION(...)  __attribute__((release_capability(__VA_ARGS__)))
 #define SHARED_UNLOCK_FUNCTION(...)     __attribute__((release_shared_capability(__VA_ARGS__)))
-#define GUARDED_BY(x)                   __attribute__((guarded_by(x)))
-#define PT_GUARDED_BY(x)                __attribute__((pt_guarded_by(x)))
+#define GUARDED_BY(...)                 __attribute__((guarded_by(__VA_ARGS__)))
+#define GUARDED_BY_ANY(...)             __attribute__((guarded_by_any(__VA_ARGS__)))
+#define PT_GUARDED_BY(...)              __attribute__((pt_guarded_by(__VA_ARGS__)))
+#define PT_GUARDED_BY_ANY(...)          __attribute__((pt_guarded_by_any(__VA_ARGS__)))
 
 // Common
 #define REENTRANT_CAPABILITY            __attribute__((reentrant_capability))